+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Helper to get open port.
 - Manager to control temp files.
 - Code to help with creating unittests that run the same tests on multiple types.
+- Function to help use tests that can either test all classes or just the "latest" one, depending on the setting of an 
+environment variable.
 
 ## 1.2.0 - 2016-09-19
 ### Added